review comment fixes

+++ core/src/main/java/org/apache/carbondata/core/util/CarbonMetadataUtil.java
@@ -407,10 +407,10 @@ public class CarbonMetadataUtil {
         new BlockletInfo3(blockletInfo.getNumberOfRows(), dimensionChunkOffsets,
             dimensionChunksLength, blockletInfo.getDimensionOffset(),
             blockletInfo.getMeasureOffsets(), blockletInfo.getNumberOfPages());
-    List<Short> rowsPerPage = new ArrayList<>();
+    List<Integer> rowsPerPage = new ArrayList<>();
     if (null != blockletInfo.getNumberOfRowsPerPage()) {
       for (int i = 0; i < blockletInfo.getNumberOfRowsPerPage().length; i++) {
-        rowsPerPage.add((short) blockletInfo.getNumberOfRowsPerPage()[i]);
+        rowsPerPage.add(blockletInfo.getNumberOfRowsPerPage()[i]);
       }
       blockletInfo3.setRow_count_in_page(rowsPerPage);
     }
